bug: Fixed Cannon Cast crashing game through stack overflow

+++ common/src/main/java/rbasamoyai/createbigcannons/crafting/casting/AbstractCannonCastBlockEntity.java
@@ -441,7 +441,8 @@ public abstract class AbstractCannonCastBlockEntity extends SmartBlockEntity imp
 		if (this.canRenderCastModel() && this.castShape != null) {
 			this.onDestroyCenterCast();
 		} else if (this.level.getBlockEntity(this.getCenterBlock()) instanceof AbstractCannonCastBlockEntity otherCast) {
-			otherCast.destroyCastMultiblockAtLayer();
+			if (otherCast.canRenderCastModel() && otherCast.castShape != null)
+				otherCast.destroyCastMultiblockAtLayer();
 		}
 	}
